// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        SessionRead: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Student Id
             * Format: uuid
             */
            student_id: string;
            /**
             * Workout Id
             * Format: uuid
             */
            workout_id: string;
            /** Workout Name */
            workout_name: string;
            /**
             * Started At
             * Format: date-time
             */
            started_at: string;
            /** Completed At */
            completed_at: string | null;
            /** Notes */
            notes: string | null;
            /** Exercises */
            exercises: components["schemas"]["ExecutionRead"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}